+++ SOSAuto_BACKEND/src/main/java/com/sosauto_backend/respository/PersonneRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonneRepository extends JpaRepository<Personne, Integer> {
+public interface PersonneRepository extends JpaRepository<Personne, Long> {
 
 }
